Move packages that should not be exposed internal (#10)
Move several packages that should not be exposed to the "itnernal" directory.
This is an effort to consolidate packages that are strictly internally
facing into a single directory that is not exposed.  This is particularly
useful in the context that we have forked several upstream libraries in an
effort to produce a solution, and we want to hide those details from the
user.  For now, I have moved all packages to "internal", and we can later
decide which ones to expose on a case by case basis.  The idea is to expose
as little as possible while still creating a useful tool.
