+++ internal/pkg/text/template/example_test.go
