+++ internal/pkg/text/template/examplefiles_test.go
