+++ internal/pkg/text/template/examplefunc_test.go
