+++ internal/pkg/text/template/exec_test.go
