+++ internal/pkg/text/template/funcs.go
