+++ internal/pkg/text/template/helper.go
