+++ internal/pkg/text/template/template.go
@@ -5,7 +5,7 @@
 package template
 
 import (
-	"github.com/redhat-nfvpe/helm-ansible-template-exporter/pkg/text/template/parse"
+	"github.com/redhat-nfvpe/helm-ansible-template-exporter/internal/pkg/text/template/parse"
 	"reflect"
 	"sync"
 )
