+++ internal/pkg/text/template/testdata/file1.tmpl
