+++ internal/pkg/text/template/testdata/file2.tmpl
