+++ internal/pkg/text/template/testdata/tmpl1.tmpl
