+++ internal/pkg/text/template/testdata/tmpl2.tmpl
